CLIENT/requirements.txt
# Requirements automatically generated by pigar.


Pillow == 6.2.0

mss == 4.0.3

opencv_python == 4.1.2.30


requests == 2.22.0